doc-service/requirements.txt
Flask==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyNaCl==1.5.0
Werkzeug==2.1.2
bcrypt==3.2.2
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
expiringdict==1.1.4
google-api-core==2.8.1
google-auth==2.7.0
google-cloud-core==2.3.1
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
launchdarkly-server-sdk==7.4.1
paramiko==2.11.0
protobuf==3.20.1
pyRFC3339==1.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
rsa==4.8
semver==2.13.0
six==1.16.0
urllib3==1.26.9
zipp==3.8.0
git+https://github.com/daxiom/service-runner.py.git@0.0.8
git+https://github.com/daxiom/simple-cloudevent.py.git@0.0.2
git+https://github.com/bcgov/registries-search.git#egg=search_api&subdirectory=search-api
# needed for search-api
git+https://github.com/bcgov/registry-schemas.git@1.5.3#egg=registry_schemas
